requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.19
djangorestframework==3.14.0
drf-spectacular==0.26.2
et-xmlfile==1.1.0
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
kombu==5.3.1
lxml==4.9.2
MarkupPy==1.14
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==1.3.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.39
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
XlsxWriter==3.1.1
xlwt==1.3.0
zipp==3.15.0
